Split shell buffer to separate file
